[add] parse files and headers

+++ include/vector.h
@@ -6,7 +6,7 @@
 /*   By: jaejilee <jaejilee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/17 16:18:28 by jaejilee          #+#    #+#             */
-/*   Updated: 2024/02/17 16:18:46 by jaejilee         ###   ########.fr       */
+/*   Updated: 2024/02/20 11:22:14 by jaejilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 
 typedef struct s_vector
 {
-	int	x;
-	int	y;
-	int	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vector;
 
-typedef t_vector t_point;
+typedef t_vector	t_point;
 
 #endif
